requirements.txt
duckduckgo-search 
streamlit 
groq 
google-search-results
python-dotenv